Major changes have been implemented in this commit. Please read the entire message for further information.
Firstly, a typedef of a Node structure has been introduced in the Node.h file, allowing us to use these structures everywhere without the prefix of 'struct.' Another important change is that the LinkedList header has been separated into an interface and implementation file. This is for modularity. Additionally, a makefile has been created to allow for the easy use of compiling/cleaning the files.